Fixed debug tests in light of addressing changes

+++ receivers/native/generic-core-2/src/main/c++/zcode/modules/core/ZcodeCapabilitiesCommand.hpp
@@ -30,9 +30,9 @@ public:
         uint8_t capabilities = 0x01;
         out->writeField8('C', capabilities);
         out->writeField32('N', (uint32_t) ZP::fieldNum);
-        uint16_t moduleName[slot.getZcode()->getModuleNumber()];
+        uint8_t moduleName[slot.getZcode()->getModuleNumber()];
         for (uint8_t i = 0; i < slot.getZcode()->getModuleNumber(); ++i) {
-            moduleName[i] = slot.getZcode()->getModules()[i]->moduleId;
+            moduleName[i] = (uint8_t) slot.getZcode()->getModules()[i]->moduleId;
         }
         out->writeBigHexField((uint8_t*) moduleName, slot.getZcode()->getModuleNumber());
     }
